refactor(harness): remove superfluous for loop in watcher
We know that change log events are emitted in batches, because they're
inserted in bulk in transactions. So the events we received from the
watcher changes channel must represent at least one full transactions
worth.

Originally, the harness would poll, timing out after LongWait, until it
had received enough events to satisfy the assertion.

This greediness lead to problems.

Instead, we should design our tests and checks with the knowledge that
we know change stream events are sent in large batches. This isn't
ideal, but is an improvement.