fix(deploy): fix a bug in deploy with --bind flag
In the apiserver, the deploy facade would map our endpoint binding space
names to space ids. This was an accommodation for Mongo.

However, the application service expects endpoint bindings with space
names, not ids. This leads to a failure to deploy with bindings other
than the default.

Fortunately, we no longer need to write our endpoint bindings to Mongo,
as we are very close to fully cutting over to DQLite.

So drop the writing of endpoint bindings to Mongo, and drop the
translation to space ids.