feat: removes pubsub from everywhere but state
This removes pubsub from the apiserver, and workers. The only place
that is left, is the underlying watcher implementations in state.
Without breaking all the watchers, this needs to be left as it's the
distribution method for all watchers.

Once state is gone and all watchers are removed, pubsub is then
gone forever.