feat: track controller target version in controller DDL
This commit adds support to the controller DDL for tracking
the target version of controllers in the cluster. This work
was introduced so that model upgrade and creation logic can
stop special casing the controller model for target versions.

This also fixes the issue where the current agent binary
version is used to work out what the controller version is.
The problem with this approach is that the current agent binary
may be lagging behind the desired version for the controller.

This work now exists as reference point to ask the question
what version is the controller running. The caller now never has
to special case the controller model or reflect the current
agent binary.