feat!: remove the instancemutater worker
This patch removes the lxd profile watching and setting feature by
removing the instancemutater worker.
The whole lxd profile related features were already broken by this point
in main. The new modelling in the application and machine domain mean
that this feature can be reimplemented in a different manner in the
future.
The facade endpoints were left and return a not implemented error.