feat: ensure that discovered IP addresses include mask suffix
When converting wire params to the domain types for network config, IP
addresses now include the subnet mask suffix determined from the
incoming address subnet.

There is a known redundancy here. The format we're converting to using
the CIDR is how we read it on the machine. Changing that aspect is out
of scope for thie exercise.