feat: update watcher mapper type
This updates the mapper type in the notify and namespace watcher.
Instead of `A -> A` which was intended to create composable mappers
(functors essentially), it creates dissonance with the inital query,
which can only return a slice of strings. It was proposed that the
initial query could return a slice of change events. I didn't want
that to happen, as it means the initial query can give the illusion
of sourcing the events from the change stream. This isn't the case,
so we need to not open that avenue.

The change is simple, map from `[]ChangeEvent -> []string`, which is
what ultimately the watchers are doing. It prevents composibility, but
nobody is using the mappers for that.

There is a fix in the lifewatcher to ensure ordering, and we should
have a test in the harness that any events that are perserved are in
the same order that they where emitted.