feat: remove units when removing applications
We also need to schedule the removal of units when removing
applications. The application is transitioned to dying, we also
transition all the application units to dying. This will signal to
others that the units and associated things are also going away.

As part of this, we'll also trigger the unit removal logic, which will
check if the unit is the last one on the machine, it will also
transition that to dying. This is currently done in another transaction,
but will have to change and be done in the same one.

Included in this is also a short circuit for removing a charm if nobody
is using it. This isn't implemented yet, that will also come later.