style: rename controller version getter
This commit renames the state func for getting the controller version
from GetControllerVersion to GetTargetControllerVersion. Target was
added to help clarify what the value is. Without target the connotation
exists that it is the version of this controller or a controller which
may not be correct.

This then aligns with the database schema.