chore: change state name for setting controller version
Changes the state method for setting the clusters controller version
from SetControllerVersion to SetControllerTargetVersion. Without target
the connotation exists that the version of this controller is being
changed and that the affects may be instance.

By adding target it is much more explicit what the operation is changing
and that it affects all controllers of the cluster. The wording also
aligns with the underlying state.