feat: set controller target version on bootstrap
This commit extends the controller config bootstrap function to seed the
target version of the controller on bootstrap. This commit made the
choice to place this in the same location where the controller's
uuid is set. This was done to make sure RI constraints can be followed.

Upon successful bootstrap the controller's target version will be
tracked as a first class entity.