fix: include architecture and base in machine/unit metrics
Updates our existing metrics collection around how many machines and units
are running in a model/controller to include their base and architecture
information.
We don't need to include "series" like we have for other metrics, because
we've never tracked series for this data. This does increase cardinality
slightly, but only by a relatively bounded number (eg ~5 values for base,
4 values for arch.)