feat: ensure we delete any charm information
We should ensure that we delete the charm because there is a unique
index on the charm name. If that wasn't the case, we could in theory
not have to worry too much about deleting a charm. Unfortunately, that's
not the case and we need to delete as much of it as possible.