feat: support token auth for migration
When a model is migrated, controller<->controller authentication is required. Traditionally this is username/password or macaroons, however JIMM supports neither. Instead, a token (specifically a JWT) can be sent to JIMM and we include logic that when a token is provided, to use the appropriate login provider so that the correct API request is made.