feat: record machine is controller
When creating an application, we want to record if the machine is
a controller or not. We don't allow modification of the controller
table, you're either a controller or not.

We also know exactly where the machine is created to become a controller
machine. Either it's during bootstrap or enable HA. Every other place
is for non-controller machines.