feat: minimal implemenation of environ.Networking for kubernetes
Ensure that the caas broken implements the environ.Network interface
to enable subnets to be added to the db when a model is created. All but
the Subnets method returns NotSupported at this time.

No actual subnets are being discovered at this time. A placeholder
subnet is being returned instead. Discovery will be done in the future.