style(network): inline multiline statements for readability
Before this patch, this file follows a strict rules of 80 characters max for each lines. This is not
 optimal for readability.

This patch release the constraint to a 120 characters max by line which is better, especially for
logs and errors formatting.

No functionality is changed; this patch focuses solely on improving code maintainability.