fix: add caching for subordinate charm info
When running `juju status` on a model, we want to report the charm
information (URL) for each unit. This includes subordinates. In the
status loop, we do a good job of only loading the Application
information once, and caching it for all of the units of that
application. However, when we hit the subordinates, we reload the charm
information over and over again for each subordinate unit. In many of
our large deployments, we might have up to 8 subordinates for every
primary unit (landscape, livepatch, telegraf, etc).

In testing with a contrived scenario with 9 subordinates for every
primary, this changed the time for `juju status` from .25s to 0.20s by
removing the read of controller config, and then to 0.14s by caching the
subordinate's charm information.