refactor: charm uuid for inserting a new unit
When inserting a unit for an application we used to fetch the charm uuid
for the new unit inside of the func that inserts the unit record in the
database.

Now that work is underway to support storage directives for units one of
the values required when creating the unit storage directives is the
charm uuid that the charm is using.

Instead of requesting this value for a second time it makes sense to
just request it once in the caller and supply it downwards to the
individual insert/create funcs.