feat: wire the caas application provisioner
This patch removes all usages of application/units but one
(`StorageConstraints()`) that will be wired when finishing the
implementation of storage in dqlite.

The result from the `ClearApplicationsResources` endpoint is
NotImplemented for every entity. This is consistent with the current
implementation since the `ClearResources` method for Application was
shimmed to return that same error.