feat: controller application
It's now possilble to know if any thing associated with the controller
application is a controller entity. For example, via a join it's
possible to know if the machine is a controller machine, or even a unit.
By moving the flag further upstream it taints the whole of the system.