feat: allow marking a unit as dead
This allows the marking of a unit as dead. This is required so that we
know we can remove a unit from the data i.e. the unit agent has done
with the unit and so it's safe for deletion. Although, we don't
guarantee in 3.6 that we can delete a unit from mongo if it's dead,
we seem to allow it to be removed anyway. Dead and NotFound might mean
the same thing ultimately.