feat: prevent deleting an alive unit
A unit MUST be already in a dying or dead state to be deleted. You can
not transition from alive to delete and miss out the lifecycle. I would
like to also ensure dying occurs, but we currently allow that in the
old model, so we should allow that in the current one.