feat: introduce new unit storage directive create method
This commit adds to the application state layer the ability to create
storage directives for a unit. This will eventually be used when
creating new units to drive storage instance creation.

Missing from this commit is the actual setting of either the store pool
or type. Somehow this implementation detail has leaked into the state
layer when it should have never got this far.