feat: introduce new type for creating unit storage directives
To properly create unit storage directives a new type is required to
represent the information that should be interested for a unit storage
directive.

This commit brings in the new type and makes an explicit attempt to
start separating a unit's directive type from that of a provider or
pool. This is a modeling concern of the facade layer that we have
allowed to traverse down to the application state layer.