refactor: re-implement storage name in storage domain
This commit re-implements storage name type in the storage domain.
Because the type was only introduced recently has little to no
dependencies it is safe to move down into the domain where it should
actually live.

This was done so the application domain doesn't have to import from so
many places to define storage directive types.