feat: introduce application storage directive type
This commit introduces a new type to the application domain to represent
storage directives for an application. Specifically the intended design
goal of this type is to be used during application deployment.

Previously this domain was using the internal storage directive type
that exists for Juju. The problem with this type is it has a very weak
contract around what the value of the pool stands for.

Sometimes it means that the value can be the name of a named pool and
other times it means that it could just refer to a provider. It is
unclear and worse it gives the impression that a specific code path
might support other value other than a storage pool.

At the service layer down we should not be taking on cognitive burden of
what the values means as it is mostly a modeling concern of the api
layer. Instead this sets a new contract for the domain demanding strong
types and explicit intention.

This work is done to support setting storage directives for a unit.