refactor: application service for deploying
This commit refactors parts of the application service for deploying new
applications. This refactor now uses the stronger application storage
directive to clearly communicate with the state layer directive intent.

By doing this the burden is removed from the state layer for having to
understand what a value is for storage pool and instead allows the state
layer to be very procedural and not concerned with data meaning.

There does exist a todo in this commit for wiring up the store pool uuid
from the name value supplied by the caller. This is intentional for the
moment.

It is also intentional to not consider provider type with storage pool
uuid at the moment. It has not been demonstrated yet that deploying an
application supports this.