refactor: application storage pool directive insert
This change refactors the insertion of application storage pool
directives. The func now uses the new arg type that clearly defines if
it is a pool or provider that is to be used.

This allows the state layer to no apply meaning to the value anymore and
instead can worry about getting the data into the database.

This work feeds into support unit storage directives as the values will
be known to be in a ready state and can just be accepted by the unit.