refactor(migration): cut over migration pre-check to DQLite
Add a new service method, CheckAllApplicationsAndUnitsAreAlive, to
optimise the pre-check phase of the migration. This reduces some complex
logic, involving many many transactions down to a single transaction.

Wire this up to the migration pre-check.

This did, however, involve a refactor of checkRelations, which now needs
to call GetUnitNamesForApplication many times (although I did add a
cache). I'm not too worried about this because really the entire check
should be pushed into a service call.