fix: lots of Printf errors
Lots of places where we passed a non-constant string as the first
argument that utimately ends up as part of a Sprintf formatting. All of
these are potential sources of bugs, where we take a user message and
then use it as the format argument, rather than as the content. More to
come.