fix: more Sprintf issues
A lot more places where we were not being careful about our format
strings to a format function.
Several of these are quite minor, or only in the test suite, but some of
these represent real places where if the content had a % in it, then we
would get undefined behavior.