chore: resolve 2.9 vs 3.6
Did a test merge of 2.9 into 3.6 to discover where my changes might
conflict with how the 1.24.4 was resolved in 3.6. There are one or two
places where I think I'm still right, but for the rest of it, resolve in
favor of 3.6, so that we can avoid conflicts.

There are still quite a few conflicts, but mostly of the form that 2.9
and 3.6 have diverged, and the file has been deleted, or the structure
is just different in 3.6.