fix: machine loopback addresses not being accounted
Some environments, like BGP, might use machine addresses assigned by BGP
instead of MAAS. These addresses are attached directly to the lo interface
but are not considered valid for the machine. As a result, it's impossible for
those environments to switch the default spacing to one that relies on lo
assignments.