fix: scale k8s units robustly with unit life + scaling state
The previous scaling algorithm suffered from race conditions
when pods introduced themselves as new units outside of strict
scaling controlled by the caasapplicationprovisioner. It also
suffered from multiple sources of truth, the juju model and
the k8s stateful set. This made scale down decisions difficult
and error prone.

This scaling algorithm is still distasteful, but now relies
on a single source of truth for scaling decisions it makes,
and, with the scaling state providing guarentees on when new
units can be expected, allows the worker to scale up and down
safely. Safe, as in, no units will lose their pod without first
transitioning to dead.