feat: propogate app storage directives to units
This commit introduces a new function to the application state layer for
fetching an applications storage directives as the argument
representation that were created with.

What this allows for is every time the state layer is attempting to add
a new unit to the model that is not apart of deploy an application the
application's storage directives can be copied to the new unit.

Units when being created follow the storage directives of their paranent
application.

Further to this commit the func is wired up to all places where units
are created to make sure that the storage directives are properly
initialised.