chore: commenting out storage creation for migration
When importing applications and units we did have code that would try to
make storage instances for each units. The code that creates the storage
instances is not correct and the migration process is currently not
setting the storage import data.

Another task needs to be created to properly establish the work here
that needs to be done.