feat: record controller in auth stack
If the machine is a controller during the auth, then record it for the
lifecycle of the websocket connection. The machine will never be not
a controller machine, after becoming a controller machine. Thus the
only way this can be invalid is if the machine is removed. This will
be noticed in a different way.