refactor: move machine placement to machine domain
This patch moves the machine placement methods to the machine domain,
but making the needed methods public so they can be called by the
application domain as well.

This was the correct place to be in the first place, and since the
application domain needs some of this functionality in the state layer,
we had to make the methods public.