feat: allow the controller uniter to come up
The uniter for the controller doesn't come up unless **all** watchers
are available. This personally is a bit much. There should be a list
of watchers that are a must and then some are optional.

The solution is to just return a TODO watcher until we can wire up
actions.