feat: allow unsubscribing during dispatch
This is a funny one, because we've always said that serializability of
subscriptions and dispatching is actually a really nice model. Except
that it's not performant enough. This model allows you to subscribe
and unsubscribe whilst processing changes. The downside is that you
can also unsubscribe during a dispatch phase. This is problematic
as you might also not have the other side of the channel reading the
information. In this case, we continue on and unsubscribe them.

This needs some serious thought, we're loosing a very nice property
here.