refactor: remove storage and storage pool kind support from args
This commit does three tasks. It stops calling and removes the insert
storage and storage attachment methods that creating a unit previously
called.

These functions are based off on a wrong premise about how unit storage
needs to be created. Specifically it doesn't take into account shared
attachements or attachments made with add unit.

It also requires the state layer to take a mapping of support storage
types for each provider. This is incorrect because the check can be done
at the service layer and the charm storage declaration is what should be
driving the type of storage instance that is made.

Because of these two removals we do not need to pass down to the state
layer anymore a map of storage providers nor do we have to define
storage to create. We can just speak in terms of directives.