refactor: add storage directive overrides to add application
The add application args supplied to the application domain service now
includes a new member to specify storage directive overrides for the
application.

This allows the caller and the user to defines which parts of a charm
storage they want to change.

This change was introduced so that storage can be removed from the same
add application args over the next few commits. The reason for doing
this is the caller should be explicit about the overrides. At the moment
this isn't possible as we take the pool value to mean either the name of
a storage pool or a provider.

This is a modeling concern of the facade that we don't want to leak
further down into the domain.

By being explicit about intent the service layer can now stop trying to
infer meaning and just be explicit about direction. This will clean up a
lot of code and reduce the likelihood of subtle bugs over time.