feat(network): subnet management during address merging
This patch introduces subnet management in `domain/network/state/linklayer_merge.go`. It provides
functionalities to rematch IPs with their expected subnet from Provider information.

Basically, it will try to reassign the subnet UUID for an IP address to a matching subnet, either
because:
- We got a provider subnet ID which matches to a known subnet
- We haven't a provider subnet ID, but the IP is linked to a placeholder subnet in Alpha space with
  one ip (ipv4/32 or ipv6/128) and we found a unique subnet that contains this address.

If an IP is rematched from a placeholder subnet, the placeholder subnet is cleaned up from the DB.

- Added `subnetToUpdate` in `mergeAddressesChanges` to track subnet updates.
- Implemented methods like `updateSubnets`, `updateSubnetFromProviderID`, and
  `updateSubnetByRematch` to manage subnet updates and rematches effectively.
- Added `cleanupUniqueAddressOrphanSubnets` to identify and remove orphan subnets efficiently.
- Updated SQL queries and logic to handle `subnet_uuid` changes in `ip_address`.