fix(network): fixes non derministic subnet matching
This patch fixes an undeterministic resolve when trying to match an IP to a subnet:

In the case of several subnet with different CIDR matches the same IP (for instance, if several
encapsulated subnets), the selected subnet would be undeterministic since the resolution rely on a
map.

To keep the same spirit of the actual implementation which accepts that a same CIDR can be linked
 to several subnet UUID, encapsulated subnets are treated the same way: if several subnets match an
 IP (either because they have the same CIDR or because they are encapsulated), it causes a "no
 error no match" results.

### Key changes:
- Updated `findSubnetUUIDForIP` in `types.go` to collect all matches and return the first UUID
when there's only one match. Improved error and edge case handling.
- Introduced `filter` helper function in `linklayer.go` for reusable slice filtering logic.