feat: tombstone subscriptions
To give back better semantics for how the event multiplexer works during
unsubscribing, this tombstones subscriptions and only removes them after
all changes are dispatched.

This prevents any subscriptions being removed during a dispatch, but
it means it's hard to test. If we're happy with this, another round
of changes to the tests will have to be done.