feat: allow a model worker agent to access model
This is insane! We first try and login in to the model using the
request model uuid, if that doesn't work _and_ it's a machine agent
that isn't in a controller model then check the credentials again. This
is fine, except that this information is knowable.

Also this code is just terrible spaghetti code, that has grown with
know sense of design. We're mixing both authn and authz all over the
place. I'm very disapointed, this should be an easy change, but it
isn't.