feat: add GetAllAPIAddressesWithScopeForAgents
Replaces APIHostPortsForAgents. A new, fit for purpose return type is
used, controllernode APIAddresses, including only the required data.

Scope is added to the controller_api_address table to be used with
APIHostPortsForAgents. The type is not validated here, but rather where
it was originally set as a host ip_address in the model's database.
SetAPIAddresses was updated to include the value.

Scope will be used to with PrioritizedForScope by the callers of
GetAllAPIAddressesWithScopeForAgents.