feat: add methods to get all api addresses for agents
GetAllAPIAddressesWithScopeForAgents returns an prioritzed slice of API
addresses for agents. The list is prioritized based on scope with
preference to cloud-local addresses.

GetAllNoProxyAPIAddressesForAgents returns a sorted comma separated
string of ip addresses indended to be used for no-proxy settings.

These replace calls to APIHostPortsForAgents.

Scope is added to the controller_api_address table to be used with
APIHostPortsForAgents. The type is not validated here, but rather where
it was originally set as a host ip_address in the model's database.
SetAPIAddresses was updated to include the value.

Functionality from the core network package has been copied over to the
domain to faciliate prioritization of the API addressers including
APIHostPortsToNoProxyString, PrioritizedForScope and ScopeMatchCloudLocal.