refactor: introduces storage directive overrides
This commit introduces a new structure to the add application args. The
new structure allows the caller to supply a list of storage directive
overrides for the application as optional values.

Before this we were using the internal storage directive type to
represent this information. This struct doesn't accurately convey what
was set by a caller and can result in ambiguity in the data. It also
propogates the weak meaning of what a pool is.

The new type forces the facade to be explicit about the values and
convey clear intent.