refactor: storage directive making in application domain
Storage directives relate directly to the application domain and not the
storage domain. They are product of application deployment. Because of
this the logic around constructing a set of storage directives has been
moved back into the application domain.

In doing this the logic for constructing the set of storage directives
is now agnostic of the model type and doesn't try to infer meaning from
user values.