feat: check charm storage requirements
This commit adds a check to validate the storage requirements of a
charm. We want to do this to make sure that a charm's requirements are
in the world of feasible.

Previously this check was done on the storage directives made for the
application. By pushing it higher and back on to the charm it is the
charm that is not passing validation and instead of implying there is a
problem with the storage directives for the deployment.