feat: implements a validator for storage overrides
This commit brings in a new validator for checking the storage directive
overrides that have been supplied by the caller when deploying a new
application.

Essentially this is validating the user input when creating
applications. The values the user supplies are easier to check at the
start then trying to check all directives at the end. We know the
defaults code for storage directive is safe and so it doesn't make sense
to run the validators over the full set.