fix: kill subscription instead of calling unsubscribe
This fixes a deadlock that occurs when subscription dispatch times out.
At this point we are entrant into a case in the multiplexer's event
loop, but we call the subscription's Unsubscribe method, which attempts
to dispatch to the same loop.

Instead, we kill the subscription and rely on the deferred Unsubscribe
call from the subscriber.