feat: add machine create args
This patch refactors the machine domain to use arguments for machine
creation instead of receiving individual parameters.

Also, and mainly, we stop receiving the machine name and uuid as
parameters in the state layer. Instead we create the machine with the
correct name (from the machine sequence) and the associated net node is
created in the same transaction.