feat: default implementation of storage provider validator
To validate application caller storage directive overrides we need to be
able to ask questions of the storage provider. Previous commits had
defined the interface needed for asking questions.

This commit provides a default implementation in the service layer.
Still missing is the state functions required and wiring up to the
application service.