refactor: remove eventmultiplexer unsubscribe function
We no longer rely on unsubscribe to remove subscriptions from the
multiplexer loop. Instead, the only action to take is killing the
subscription.

The multiplexer now simply removes such subscriptions on each pass
through the loop. The subscription Unsubscribe function is now a no-op
and will be cleaned up in a commit to follow.