refactor: remove storage directive validation from state
This commit removes the last vestiges of storage directive validation
from the application state layer. Validation of storage directives
should now be done exclusively at the service layer.

As part of this commit I have had to comment out some test code that
asserts filesystem and volumes. This will come back in as a seperate
commit later when the questions can properly be made.