refactor: storage suites in application domain
The testing suites for storage in the application domain are pure
insanity. They are not calling the correct setup methods and doing busy
work that should be a no no in testing.

I have managed to remove the need for a model out of some the suites but
more work will need to be done down the line.