test: add tests for getting pool provider type
This commit adds tests to the state layer of the application domain for
asserting the GetProviderTypeOfPool. This state layer function is used
when validating the storage pool provided by a caller as a directive
override.

This is common when users want to deploy an application to a model but
wish to change the default storage used in the model.