refactor: remove storage registry getter
This commit removes the no longer used storage registry getter from the
application domain services. The reason for this having existed was to
build a map of storage providers within the model so that validation can
be performed. The providers also helped influence what type of storage
instance was created.

The information is no longer needed as the validation of a provider type
is now abstracted behind an interface allowing the value to be easily
mocked. We also don't require the registry in the state layer to make
storage instances anymore as this data can be driven off of the
directive.