feat(caas/kubernetes): support --attach-storage option for deploy application on k8s
Add support for deploying with the --attach-storage option on Kubernetes
clouds.

This change updates multiple API schemas to ensure the CAAS provider can
access
the attached storage information. The CAAS provider will create a
PersistentVolumeClaim with the specified volume name in its spec to
ensure it
binds to the target PersistentVolume.