feat: remove machine is manager
The machine is manager is used to check if the machine is the
controller application machine in the controller model. This is such a
worthless check most of the time, because we can omit the call just
because the model uuid isn't the same as the controller model uuid.
Thus a lot of queries can just be removed.

Additionally, I've also removed the set machine addresses so it becomes
a no-op.