test: update the controller tests
We don't have a way to represent the controllers as part of a
leadership. The peergrouper gave that information to us. It's possible
to see what is part of the database quorum, but it's not a show blocker
for now.

I think a better alternative is to expose specific database commands,
so that it allows you to modify and view in a juju way.