feat: remove jobs from machine and state
This is a ruthless PR in that it rips out jobs completely and be done
with it. We can get the API responses to return the right response if
it's a controller. All machines can host units, but only if it's a
controller can it manage models. This was over engineered and was
probably intended for something else. Anyway, that's gone now.

There will be a lot of broken tests, which we'll have to deal with
later.