refactor: remove storage watchers
This commit removes the storage watchers that have already been
established. The reason for this is the watchers we require for storage
are all around provisioning of storage.

Provisioning watchers are very different to the general triggers that
were setup and require a different approach. Because these watchers are
not used and will most likely not be used we remove them now.

One watcher that is being kept is the block device watcher. This was
lumped under the triggers for storage and has been moved to its own
file.