feat: storage entity provisioning triggers
This commit introduces a set of new triggers for storage entities. It
allows a watcher to be created that can inform the caller when
provisioning work exists for a storage entity in the Juju model.

Storage is provisioned from either the model or from a machine. We
currently don't have a clear way to represent this in the DDL for the
model. To make this commit work and to keep writing tests a throw away
column has been added to the DDL.

This will be removed in a later commit when rebased on top of the real
DDL change required.

Custom triggers needed to be created because for each storage entity
change a dance needs to be done to work out if the change is relevant
for the namespace and if it relates to a watched entity.