refactor: updates the triggers for storage machine provisioning
Updates the provisioning triggers for volume and filesystem when they
are machine provisioning. The delete trigger now makes an event for the
last attachment being deleted for the entity.

This reason for thinking this way is if these entities are machine
provisioned then we can't tell what machine provisions the entity when
no attachments exist.

By firing this informaton on the last attachment we can still work this
information out.

Not 100% sure this is correct for the moment but it gets us a start.