feat: machine provision watcher mapper
This commit introduces a new generic mapper for watchers that care about
machine provisioned storage entities. Machine provision watchers care
about watching for changes that involve their net node or machine uuid
but the caller that is using the watcher wants to know the id value of
the storage entity that has changed.

We know that when the watcher fires a storage entity has changed. This
mapper just does the leg work to figure out what has changed and then
report back a new set of changes for the storage entity ids which have
changed.