fix: potential state deviation of entity life mapping
Because of how our watcher infrastructure works Harry correctly noticed
that between the initial query that is done for a watcher and
establishment of the mappers in memory state the possibility exists for
the two states to differ.

In simple terms this means the thing that is watching may never see a
dead or dying change for something that it thinks it knows about in the
system. The opposite is also true that we may tell a watched entity
about something that had until that point never heard about.

The fix for this is to provide the helper functions for setting up the
initial query that should be used with the mapper. This way both pieces
of the watcher puzzle are in sync.

To support the users of this mapper and initial query a new make
function was introduced that provides the caller with both watcher
pieces initialised on the same data.