fix: prevent changestream restarting
The changestream would attempt to restart the worker if the namespace
for the database was not valid. Essentially, if the call to GetDB
returns an error that should prevent a restart, then we need to handle
it in the ShouldRestart function. If the restart happens during the
start of the worker, and there is a subsequent call to get the change
stream from the runner, it will keep waiting until the abort channel
is called. This is problematic. We should audit all runner start
worker methods to ensure that any error returned back from the function
is handled correctly.

At the very least, all runners should be correctly wired up to the
loggers.