fix(migration): precheck used wrong services
When the migration precheck was initially cut over, some mistakes were
made in the service method calls.

The migration precheck runs various assertions against both the model
being migrated and the controller model. However, we only used services
for the model being migrated.

What we need instead is to inject domain service getters into the
prechecker, for the prechecker to then use to make it's checks against
the correct models.