fix: return all NICs indexed by space for DevicesToBridge
This is done because we may get a VLAN with a parent that has no
addresses, but that parent would not be found is we only select get the
list of devices in the spaces we require.

We already discount such devices when searching for pertinent NICs, so
the change is minor.