fix: resolve flakey tests due to maps
Update GetAllAPIAddressesWithScopeForAgents and
GetAllAPIAddressesWithScopeForClients to return
[][]controllernode.APIAddress rather than a map. This helps to improve
the order of results to allow for consistent test results.

However, the slice of data from the database is in a consistent order
when a slice of data is added at the same time. Thus SameContents must
still be used at a certain part. Luckily the state method results are
ordered in the service.

This change also ensures that the addresses are ordered by controller
machine ID.